chore: split ci jobs into two (#415)
* chore: split ci jobs into two

* chore: linted

+++ pallets/bridge-registry/src/weights.rs
@@ -56,13 +56,13 @@ impl<T: frame_system::Config> WeightInfo for WebbWeight<T> {
 	// Storage: BridgeRegistry Bridges (r:1 w:1)
 	fn set_metadata() -> Weight {
 		Weight::from_ref_time(6_000_000)
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().reads(1_u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	// Storage: BridgeRegistry ResourceToBridgeIndex (r:0 w:1)
 	fn force_reset_indices() -> Weight {
 		Weight::from_ref_time(862_000_000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 }
 
@@ -71,12 +71,12 @@ impl WeightInfo for () {
 	// Storage: BridgeRegistry Bridges (r:1 w:1)
 	fn set_metadata() -> Weight {
 		Weight::from_ref_time(6_000_000)
-			.saturating_add(RocksDbWeight::get().reads(1 as u64))
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().reads(1_u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	// Storage: BridgeRegistry ResourceToBridgeIndex (r:0 w:1)
 	fn force_reset_indices() -> Weight {
 		Weight::from_ref_time(862_000_000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 }
